Reformatted file structure for new rampup versions
