--- targets/phase1/main.cpp
@@ -1,19 +1,0 @@
-/**
- * This is a simple example of taking user input and printing output over UART.
- */
-
-int main() {
-    // Initialize system
-
-    // Set up UART
-
-    // Declare two arrays of characters to store user input
-
-    while (1) {
-        // Read user input
-
-        // Convert the two character arrays to uint64_t
-
-        // Output the sum of the two numbers
-    }
-}
